Do not use an attribute to manage the job ID
Instead the Job id is managed internally by the orchestrator.
The contract is to export a TOSCA_JOB_ID output in the submit operation
Then this TOSCA_JOB_ID is injected as an input of run and cancel ops.
The run operation is designed to be called periodically by the
orchestrator to check the job status.
The run operation exports a TOSCA_JOB_STATUS output that could be:
* RUNNING
* COMPLETED
*FAILED

To indicated